+++ js/variables.js
@@ -1,0 +1,9 @@
+const AVATAR_MIN = 1;
+const AVATAR_MAX = 6;
+const LIKES_MIN = 15;
+const LIKES_MAX = 200;
+const COMMENTS_MIN = 0;
+const COMMENTS_MAX = 30;
+const MAX_NUMBER_PHOTO_ID = 25;
+
+export const veriables = {AVATAR_MIN, AVATAR_MAX, LIKES_MIN, LIKES_MAX, COMMENTS_MIN, COMMENTS_MAX, MAX_NUMBER_PHOTO_ID};
